fix(entance): error when moving entrance with no cave

+++ api/controllers/v1/entrance/move-to-cave.js
@@ -1,6 +1,5 @@
 const CaveService = require('../../../services/CaveService');
 const ControllerService = require('../../../services/ControllerService');
-const ErrorService = require('../../../services/ErrorService');
 const { toEntrance } = require('../../../services/mapping/converters');
 
 const { checkIfExists } = sails.helpers;
@@ -22,43 +21,36 @@ module.exports = async (req, res) => {
     });
   }
 
-  try {
-    // Remove entrance from initial cave and possibly delete initial cave
-    const entrance = await TEntrance.findOne(entranceId);
+  // Remove entrance from initial cave and possibly delete initial cave
+  const entrance = await TEntrance.findOne(entranceId);
 
+  if (entrance.cave) {
     await TCave.removeFromCollection(entrance.cave, 'entrances', [entranceId]);
     const initialCave = await TCave.findOne(entrance.cave).populate(
       'entrances'
     );
 
     if (initialCave.entrances.length === 0) {
-      await TCave.update(initialCave.id).set({
-        redirectTo: destinationCaveId,
-      });
+      await TCave.update(initialCave.id).set({ redirectTo: destinationCaveId });
       await CaveService.deleteCave(req, initialCave.id);
     }
-
-    // Add entrance to destination cave
-    await TCave.addToCollection(destinationCaveId, 'entrances', [entranceId]);
-
-    // Return populated entrance
-    const updatedEntrance =
-      await TEntrance.findOne(entranceId).populate('cave');
-    updatedEntrance.cave.entrances = await TEntrance.find().where({
-      cave: entrance.cave.id,
-    });
-    const params = {
-      controllerMethod: 'EntranceController.moveToCave',
-    };
-    return ControllerService.treatAndConvert(
-      req,
-      null,
-      updatedEntrance,
-      params,
-      res,
-      toEntrance
-    );
-  } catch (e) {
-    return ErrorService.getDefaultErrorHandler(res)(e);
   }
+
+  // Add entrance to destination cave
+  await TCave.addToCollection(destinationCaveId, 'entrances', [entranceId]);
+
+  // Return populated entrance
+  const updatedEntrance = await TEntrance.findOne(entranceId).populate('cave');
+  updatedEntrance.cave.entrances = await TEntrance.find({
+    cave: destinationCaveId,
+  });
+
+  return ControllerService.treatAndConvert(
+    req,
+    null,
+    updatedEntrance,
+    { controllerMethod: 'EntranceController.moveToCave' },
+    res,
+    toEntrance
+  );
 };
